+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.1.4](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.1.4) (2016-05-11)
+Updated AdColony library dependency to fix bugs.  
+Fixed AdColony reporting wrong ad shown state.  
+Fixed minimum interval between ads not being respected.
+
 ## [1.1.3](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.1.3) (2016-05-09)
 Fixed crash when requesting ads without a network connection.